[3.2] create spring-webmvc-3.1-core project
