+++ settings.gradle
@@ -153,6 +153,7 @@ include ':instrumentation-core:aws-sdk:aws-sdk-2.2'
 include ':instrumentation-core:reactor-3.1'
 include ':instrumentation-core:servlet'
 include ':instrumentation-core:spring:spring-web-3.1'
+include ':instrumentation-core:spring:spring-webmvc-3.1-core'
 
 // exporter adapters
 include ":auto-exporters"
